app/src/androidTest/java/com/ayevbeosa/testingworkmanager/BackupWorkerTest.kt
package com.ayevbeosa.testingworkmanager

import android.content.Context
import androidx.test.core.app.ApplicationProvider
import androidx.test.ext.junit.runners.AndroidJUnit4
import androidx.work.ListenableWorker
import androidx.work.testing.TestListenableWorkerBuilder
import androidx.work.workDataOf
import com.ayevbeosa.testingworkmanager.work.BackupWorker
import kotlinx.coroutines.runBlocking
import org.hamcrest.MatcherAssert.assertThat
import org.hamcrest.Matchers.`is`
import org.junit.Before
import org.junit.Test
import org.junit.runner.RunWith

@RunWith(AndroidJUnit4::class)
class BackupWorkerTest {
    private lateinit var context: Context

    @Before
    fun setup() {
        context = ApplicationProvider.getApplicationContext()
    }

    @Test
    fun testBackupWorker() {
        val inputData = workDataOf("ResponseCode" to 200)
        val worker = TestListenableWorkerBuilder<BackupWorker>(context, inputData).build()
        runBlocking {
            val result = worker.doWork()
            assertThat(result, `is`(ListenableWorker.Result.success()))
        }
    }

    @Test
    fun testBackupWorkerNetworkError() {
        val inputData = workDataOf("ResponseCode" to 404)
        val worker = TestListenableWorkerBuilder<BackupWorker>(context, inputData).build()
        runBlocking {
            val result = worker.doWork()
            assertThat(result, `is`(ListenableWorker.Result.success()))
        }
    }
}